move tests to right place, combine duplicate tests
